Fix fatal error when top and bottom VCs are nil
The flow of logic currently checks that both top and bottom aren't nil, then assumes that all is well if top isn't nil. This causes a fatal crash if top VC and bottom VC are both both nil. Ideally the nil check should be a parent if statement here so that the statements don't have to be continually checked (or use guard).